refactor: :recycle: resume projects app router로 이동
