+++ src/components/ResumeComponent/ResumeProjects/ProjectDigestComponent.tsx
@@ -1,9 +1,6 @@
-import { css } from '@emotion/react';
-
 import { ShareIcon } from '@/assets/svgs';
 
-import { COLORS } from '@/styles/colors';
-import { SIZES } from '@/styles/sizes';
+import styles from './resumeProjects.module.scss';
 
 interface Props {
   projectKey: string;
@@ -15,7 +12,7 @@ interface Props {
 
 export const ProjectDigestComponent = ({ projectKey, digest }: Props) => {
   return (
-    <ul css={listWrapperStyle}>
+    <ul className={styles.listWrapper}>
       {digest.map(({ description, link }, index) => (
         <li key={`${projectKey}-${index}`}>
           <span>{description}</span>
@@ -29,35 +26,3 @@ export const ProjectDigestComponent = ({ projectKey, digest }: Props) => {
     </ul>
   );
 };
-
-const listWrapperStyle = css({
-  paddingLeft: 20,
-
-  li: {
-    listStyle: 'square',
-    marginBottom: 10,
-  },
-
-  'li:last-child': {
-    marginBottom: 0,
-  },
-
-  span: {
-    color: COLORS.BLACK,
-    lineHeight: `${SIZES.FONT_MS + 4}px`,
-    fontSize: SIZES.FONT_MS,
-  },
-
-  a: {
-    width: 15,
-    height: 15,
-    display: 'inline-block',
-    marginLeft: 5,
-  },
-
-  svg: {
-    width: 15,
-    height: 15,
-    fill: COLORS.GRAY9,
-  },
-});
