+++ src/components/ResumeComponent/ResumeProjects/index.tsx
@@ -1,8 +1,8 @@
-import { css } from '@emotion/react';
-
 import { ResumeProjectListType } from '@/types/resumeData';
 import { ProjectComponent } from './ProjectComponent';
 import { Section } from '../Section';
+
+import styles from './resumeProjects.module.scss';
 
 interface Props {
   projectData: ResumeProjectListType;
@@ -12,7 +12,7 @@ export const ResumeProjects = ({ projectData }: Props) => {
   const projectKeys = Object.keys(projectData);
   return (
     <Section title='프로젝트'>
-      <ul css={projectsWrapper}>
+      <ul className={styles.projectsWrapper}>
         {projectKeys.map(
           (projectKey, index) =>
             projectData[projectKey].digest && (
@@ -23,7 +23,3 @@ export const ResumeProjects = ({ projectData }: Props) => {
     </Section>
   );
 };
-
-const projectsWrapper = css({
-  width: '100%',
-});
