+++ src/components/ResumeComponent/ResumeProjects/resumeProjects.module.scss
@@ -1,0 +1,100 @@
+@use '@/styles/constants/colors';
+@use '@/styles/constants/sizes';
+
+.projects {
+  width: 100%;
+}
+
+.projectWrapper {
+  width: 100%;
+  display: grid;
+  grid-template-columns: 25% calc(75% - 10px);
+  gap: 10px;
+  padding: 20px 0;
+
+  &:not(:last-child) {
+    border-bottom: 1px solid colors.$GRAYE;
+  }
+  &:first-of-type {
+    padding-top: 0;
+  }
+}
+
+.iconAndImageWrapper {
+  display: flex;
+  flex-direction: column;
+
+  h3 {
+    font-size: sizes.$FONT_ML;
+    margin-bottom: 5px;
+    color: colors.$GRAY6;
+  }
+
+  & > span {
+    color: colors.$GRAYA;
+    font-size: sizes.$FONT_MS;
+    margin-bottom: 15px;
+  }
+}
+
+.linkWrapper {
+  display: flex;
+  flex-direction: row;
+
+  a {
+    width: 20px;
+    height: 20px;
+    margin-right: 10px;
+  }
+
+  svg {
+    width: 20px;
+    height: 20px;
+    fill: colors.$GRAY6;
+  }
+}
+
+.digestWrapper {
+  display: inline-block;
+  color: colors.$BLACK;
+  font-size: sizes.$FONT_MS;
+  line-height: calc(sizes.$FONT_MS + 4px);
+  margin-bottom: 10px;
+
+  & > span {
+    display: inline-block;
+    margin-bottom: 5px;
+  }
+}
+
+.listWrapper {
+  padding-left: 20px;
+
+  li {
+    list-style: square;
+    margin-bottom: 10px;
+  }
+
+  li:last-child {
+    margin-bottom: 0;
+  }
+
+  span {
+    color: colors.$BLACK;
+    line-height: calc(sizes.$FONT_MS + 4px);
+    font-size: sizes.$FONT_MS;
+  }
+
+  a {
+    width: 15px;
+    height: 15px;
+    display: inline-block;
+    margin-left: 5px;
+  }
+
+  svg {
+    width: 15px;
+    height: 15px;
+    fill: colors.$GRAY9;
+  }
+}
