+++ src/components/ResumeComponent/index.ts
@@ -1,3 +1,4 @@
 export { ResumeHeader } from './ResumeHeader';
 export { ResumeIntroduce } from './ResumeIntroduce';
 export { ResumeFootprints } from './ResumeFootprints';
+export { ResumeProjects } from './ResumeProjects';
