+++ src/components/ResumeComponent/resumeHeader.module.scss
@@ -1,6 +1,5 @@
 @use '@/styles/constants/colors';
 @use '@/styles/constants/sizes';
-@use '@/styles/responsive';
 
 .headerWrapper {
   width: 100%;
