+++ src/components/ResumeComponent/section.module.scss
@@ -1,6 +1,5 @@
 @use '@/styles/constants/colors';
 @use '@/styles/constants/sizes';
-@use '@/styles/responsive';
 
 .sectionWrapper {
   width: 100%;
