+++ src/components/index.ts
@@ -12,4 +12,4 @@ export { ProjectComponent } from './ProjectComponent';
 export { FootprintElement } from './FootprintElement';
 export { FootprintComponent } from './FootprintComponent';
 
-export { ResumeHeader, ResumeIntroduce, ResumeFootprints } from './ResumeComponent';
+export { ResumeHeader, ResumeIntroduce, ResumeProjects, ResumeFootprints } from './ResumeComponent';
